+++ tests/FreeElephants/LogTracer/SimpleTraceContextTest.php
@@ -87,4 +87,17 @@ class SimpleTraceContextTest extends TestCase
 
         $this->assertTrue($context->isInitialized());
     }
+
+    public function testPopulateWithValues(): void
+    {
+        $context = new SimpleTraceContext();
+
+        $this->assertFalse($context->isInitialized());
+
+        $context->populateWithValues('4bf92f3577b34da6a3ce929d0e0e4736', '00f067aa0ba902b7');
+
+        $this->assertTrue($context->isInitialized());
+        $this->assertSame('4bf92f3577b34da6a3ce929d0e0e4736', $context->getTraceId());
+        $this->assertSame('00f067aa0ba902b7', $context->getParentId());
+    }
 }